+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [1.10.0] - 2023-11-02
+### Added
+- The `GetCategoryMonthTotals` method, which returns the total amount summed for
+  each month for a category.
+  - This is based on `entityUUID` for a specific user.
+
 ## [1.9.0] - 2023-10-29
 ### Updated
 - Force updated the tag version.